adrift/requirements.txt
Django==5.0.6
psycopg2-binary==2.9.6
gunicorn==21.2.0
graphene-django
django-graphql-jwt==0.4.0
django-graphene-auth
django-countries
bleach==6.1.0
django-phonenumber-field[phonenumbers]
kafka-python==2.0.2
git+https://github.com/hohkokyoung/django-language-field.git